Deleted thumbnail module and move its functions to image module
