+++ back/src/main.rs
@@ -23,7 +23,6 @@ mod log;
 mod openapi;
 mod schema;
 mod storage;
-mod thumbnail;
 
 use crate::{
     auth::{AuthContext, OAuth2Clients},
